+++ app/uk/gov/hmrc/entrydeclarationintervention/controllers/api/DocumentationController.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 HM Revenue & Customs
+ * Copyright 2025 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.entrydeclarationintervention.controllers
+package uk.gov.hmrc.entrydeclarationintervention.controllers.api
 
 import controllers.Assets
 import play.api.libs.json.Json
