+++ test/uk/gov/hmrc/entrydeclarationintervention/controllers/DocumentationControllerSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 HM Revenue & Customs
+ * Copyright 2025 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,6 +29,7 @@ import play.api.test.Helpers._
 import play.api.test.{FakeRequest, Helpers, Injecting}
 import play.api.{Application, Environment, Mode}
 import uk.gov.hmrc.entrydeclarationintervention.config.MockAppConfig
+import uk.gov.hmrc.entrydeclarationintervention.controllers.api.DocumentationController
 
 class DocumentationControllerSpec extends AnyWordSpecLike with Matchers with OptionValues with MockAppConfig with Injecting with GuiceOneAppPerSuite {
   override lazy val app: Application = new GuiceApplicationBuilder()
